SR-IPv6: applied draft-08
Packets to be SRH-ified are encapsulated in an outer IPv6 header which contains
the SRH. Packets with SRH segleft == 0 and nexthdr == NEXTHDR_IPV6 are
"manually" decapsulated (i.e. network and transport offset are set w.r.t
inner header and skb is reinjected in dst_input). We cannot rely on in-place
mechanisms for inner header processing because the network stack expects a
tunnel interface to be present, i.e. a stateful tunnel, which is not scalable.
Dirty hooks for local_out packet processing have been removed. Currently, the
only way to insert an SRH in a locally generated packet is to set the correct
IPV6_RTHDR socket option.