SR-IPv6: add support for netns-wide sr6 tunnel source
Commit 3fa9878f1f47b8ce93ca2fc95ff2556b1578c6df introduced seginfo caching
in order to store the source address computed for the given destination,
because of otherwise expensive, repeated calls to ipv6_dev_get_saddr.
However, there is no particular reason that the tunnel should have a
specific interface source address. If we can define a global, loopback
or not, source address for the virtual tunnels, then we both remove the need
to call ipv6_dev_get_saddr and remove the need for seginfo caching.