sr-ipv6: remove srh offloading
When receiving a packet with an inline SRH and GRO enabled, and the host is
the final destination of the packet but the active segment is not the last
segment (i.e., the packet will jump one or more times to the looped_back
label in ipv6_srh_receive() and the DA of the packet will change),
with an inner protocol that verify checksums and support GRO (e.g., TCP),
the protocol's gro_receive function will set the skb->csum_bad = 1 because
the current pseudo-header does not match the initial/final pseudo-header as
the current DA of the packet is not the final one. The packet will be flushed
upward by GRO with csum_bad=1, causing it to fail the subsequent
skb_checksum_init by the inner protocol.

This commit removes the INET6_PROTO_GSO_EXTHDR flag from the ROUTING protocol
as no assumption can be made on such packets.